requirements.txt
aniso8601==4.0.1
appnope==0.1.0
backcall==0.1.0
Click==7.0
decorator==4.3.0
flasgger==0.9.2
Flask==1.0.2
Flask-JWT-Simple==0.0.3
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.1
grpcio==1.17.1
grpcio-tools==1.17.1
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
mistune==0.8.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.7
PyYAML==3.13
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
